// Frontend/angular-frontend/src/app/models/user.ts
export class User{

    firstName: String;
    lastName: String;
    email: String;
    password: String;
    confirmPassword: String;
    profileImage: String;
    role: String;
  

    address: {
        streetAddress: String;
        city: String;
        state:String;
        zipcode: String;
    }

}